requirements.txt
fastapi
uvicorn
pythonocc-core
